Game: Move dice roll to Game and added `tick` method
Moved the dice roll from something controlled by the caller to the Game
class. The caller shouldn't be able to "re-roll". This is now a private
property on the Game class called `__roll` and there is a getter that is
used to access it.

This means that the `roll_die` function has been moved to the Game
class as a private method called `__roll_die`.

Additionally, the `tick` method has been added to handle advancing the
game turn by turn. This avoids having to check if the game is over after
the caller has rolled their die from outside the Game class. Since we
are calculating game_over every tick, `check_game_over` has been renamed
to a private method called `__calculate_game_over` which is ran every
tick with the new roll.