Game: Add `shut_boxes` method to apply a move
The logic for validation and box closing was handled by the caller when
they really belong on the Game class.

This commit added a new method to the Game class called `shut_boxes`
which applies a move. The validation checking now occurs inside this
`shut_boxes` method. Validation errors are now raised if there is a
problem applying a move where specific errors can be handled
programatically.

This was done in an effort to encode the validation errors in an
exception rather than printing an error to the console from inside the
Game class and returning a nebulous True/False.